2009-01-05	Joel Sherrill <joel.sherrill@oarcorp.com>
	* score/include/rtems/score/object.h,
	score/src/objectallocatebyindex.c: Object index should be int.
	Fix bug when index is negative.
	* score/src/objectextendinformation.c: Do not allow maximum number
	of allocated objects to exceed maximum representable in index
	field of Object Id.
	* score/src/objectgetisr.c: Use same code that is in _Objects_Get
	to extract index field of Object Id.

+++ cpukit/score/include/rtems/score/object.h
@@ -519,7 +519,7 @@ Objects_Control *_Objects_Allocate(
  */
 Objects_Control *_Objects_Allocate_by_index(
   Objects_Information *information,
-  uint16_t             the_index,
+  int                  the_index,
   uint16_t             sizeof_control
 );
